locate: now works on modules
Since [Env.lookup_module] doesn't return any location information, but only a
path (and a module_type, but we don't care here) and there is no way to
differenciate between a path pointing to a locally defined module or to another
file, we need to remember all the locally encountered modules.

That code *needs* to be cleaned.

Also I suspect there might be a synchronisation bug (compare [State.node_at] to
[State.local_modules]).