Vim: Fix locate when editing in sub-directories
The locate call returned the file name, not a full path so the locate opens the
wrong path in Vim.

E.g. If I edit using `vim src/foo.ml` and call locate on something contained in
`src/bar.ml` then it will fail and open a buffer `./bar.ml`.