locate: better handling of includes in local buffer
Before : when we got a path, say Foo.Bar.baz, from the environment and the
location was a ghost one, we tried to look for Bar.baz in foo.cmt.

Now : we first check if Foo doesn't come from an include upper in the buffer. If
it is we stop on the include, if it isn't we have the same behavior as before.

Future plans : try and follow the include.