emacs: add option to call ocamlmerlin through opam config exec
This allows switching compilers and not having to restart Emacs to use
the new compiler.

It adds 2 new options:
- merlin-command-use-opam (boolean, default nil)
- merlin-command-opam-command (path to opam, default "opam")

It works by calling
 opam config exec "ocamlmerlin arg1 ... argn"
if merlin-command-use-opam is set.

I believe setting merlin-command-use-opam to t by default would do more
good than harm (since I believe that many end-users will probably use
merlin through opam). But I left it to nil because I was not sure. A
safe way would be to set it to t when we can detect that merlin.el was
installed through opam (but how to do that?).