README: clean up.
I removed a lot of information redundant with what can be found on the wiki,
only minimal informations are given now.
I give a link to the wiki as a "compensation".

The README is imho meant for people wishing to install merlin from sources and
start hacking it, so it still need to give some pointers for the second part.
But we'll get there in time.

Some information has been "lost" with this commit (e.g. how to install merlin
with NeoBundle for vim), I will try to readd these information on the wiki soon.