Drop text properties from command sent to merlin process.
Some of the arguments that are sent to merlin process, in particular
buffer-file-name, might contain text properties. This changes how they
are printed and that causes merlin process to choke on them which in
turn creates bunch of "spurious communication" errors in emacs. Just
dropping text properties fixes the issues.