[Merlin] Add support for SUFFIX flags.
Summary: In the `ocamlc`/`ocamlopt` compiler, there is no restriction on
which file suffixes may be used. To use a file with a non-standard
suffix, you would supply the file as either `-intf` or `-impl` (note,
you would also tell the compiler that a `-impl` file has a corresponding
interface in a particular suffix by specifying `-intf-suffix`). However,
until this diff, merlin only searches for files with the `.ml/.mli`
suffixes. This diff adds support for checking other suffixes when
locating files. Merlin's heuristic will sometimes try appending `.ml` to
the end of a module name, etc - this diff allows you to instruct it to
also try other suffixes. It works great!

Test Plan:

Reviewers:

CC: