type enclosing: don't print big modules on low verbosity
In some cases, with -short-paths printing the signature of some module can take
very long (even though the module itself contains only few entries).
This gets annoying when you don't actually want the type of the module but the
module is a parent enclosing of a thing you're trying to typecheck.

This commit is only a quick&dirty hack, but I don't have a fully satisfying
clean solution yet.