Fix #662 - TypeOf error with neovim
it looks like neovim has a different indexing than vim, that causes
vim.buffers[idx] to throw a `list index out of range` error.

the code in this pr will make it work on both vim and neovim by building
a list of buffers that match the id, rather than assuming that the
buffer id matches the bufferlist id.