locate: use loc of modules/modtypes from env when available
Unrelated to this, but the implementation should probably be changed at this
point because if people want to go to "the implementation" (and not the
interface) of something, if the lock is present in the environment they will be
taken to an interface most likely.
This will become more apparent after https://github.com/ocaml/ocaml/pull/1219 is
merged.