locate: always walk up the typedtree, regardless of whether a loc was in the env.
This is a follow up to what was said in 936f487fa65a5005d051baa963dc5902c6e484ca
We could add some short-circuiting when the user is not trying to go to an ml,
and then use the loc in the environment.
But I don't think it's worth making the code (which is already a pile of hacks)
more complex.