Add lsp frontend
This adds `ocamlmerlin-lsp` executable which implements lsp frontend for
merlin using `lsp` library introduced in the previous commit.

The implementation is simple - it translates LSP requests into merlin
new protocol query commands for the exception of
`textDocument/typeDefinition` which is implemented from scratch as
there's no analogue in existing protocol.

The executable is being installed as a part of a new `merlin-lsp.opam`
package. The reason is partly b/c `merlin-lsp.opam` requires more
3rd-party dependencies (`ppx_deriving_yojson` in particular) and partly
because the implementation is relatively untested.