Browse tree: make pat_extra siblings of pattern nodes
PR "Fix wrong type information for constructors under type coercions #1091"
did the same to expressions, but it turns out that we have the same
issue with pattern.

When doing type enclosing on `def` in:

        let def : float = float_of_int 3

The browse structure looks like:
[ core_type; core_type; pattern; value_binding; structure_item; structure ]

The two core types come from a `Tpat_constraint` which is misplaced
because of the typedtree representation.