adds module, module type, and class imenu items
These are important syntactic categories, which I am missing a lot. In fact, the modules category is the most important as it enables to quickly move between submodules and significantly reduces cognition burden during developing OCaml programs. The class category fits into the same niche, except when the OOP style is employed. There is also a constructor category, though I am not really sure if is useful, and given that it can pollute the outline I decided to leave it.

uses the correct designation for the module types

The outline protocol uses "Signature" to denote a module type (see `query_json.ml:json_of_outline`). For consistency, I also renamed all other occurnces of modtyp(e) to signature.