typing-recovery: reintroduce naive record recovery
Instead of the generic `*type-error*` ident node, we get an empty record
expression `{ }`. Which is useful for various analysis which look at the
typedtree/browsetree to decide in which context they are working.

We could make the analysis less naïve by keeping all the sub expressions
which typechecked correctly (currently, we would drop all of them if one
field is missing, or if one extra field is present).