Allow using a custom/home made dot reader
At Facebook, we are using both dune and Buck to build ocaml projects.
With the recent changes in dune's behavior w.r.t. .merlin files (no
longer generated in the source tree), we need a way to tweak merlin's
logic to suits our dev environment.

Merlin's current behavior is to look into the directory of the target ml
file for a .merlin file. If it's there, call `dot-merlin-reader`. If it
is not, tries to detect dune files, and call dune.

This commit adds a simple logic to override merlin's behavior: merlin
will first look for a `.merlin.custom` file. If it is there, it will
call `dot-merlin-custom-reader` on it, and stops there.

This reader will be provided by us (or any third party that want some
custom logic).
The name of the reader can be overriden by an environment value
`DOT_MERLIN_CUSTOM_READER`