Setup load path only for typer
The typer needs to be executed with proper settings. It is too expensive
to setup load path for reader and typer, but the reader should not
depend on it (AFAICT). By setting the load path only for the typer, we
don't pay the cost twice and are executing each part of the pipeline in
the environment they expect.