Fail gracefully when source files of preprocessed files are absent.
This commit addresses the request in
https://github.com/ocaml/merlin/pull/1219#issuecomment-830728555

It also adds a test analogous to dot-pp-dot-ml.t that builds with dune and
demonstrates that the fix works. I'm still unsure what's wrong with the original
test, but since it's hand-constructed I'm not sure that it's simulating a real
case.

Finally, it updates both tests to use different values for x in dep.ml to avoid
triggering an unrelated Merlin bug - https://github.com/ocaml/merlin/issues/1403
- which should be resolved separately.