Windows: setup ppx command redirection with CreateProcess
The code introduced in #1270 switched from using system() to
CreateProcess() in order to not show a window for each execution of a
ppx. However, it kept 1>&2 which is shell syntax, leading to errors.

With this patch, we revert to Sys.system and appropriate quoting for
Unix systems, but on Windows we wrap CreateProcess and use the same
file descriptor for stdout and stderr to have the same redirection as
with the system() call.