Implement shape-based locate
Fix locate for documentation queries

Previously the locate function would relies on a Typedtrie search and ocasionally return the documentation at the same time. With shape-based locate this is not the case anymore and we always rely on a call to `Ocamldoc.associate_comment`.

This commit also fix file_switching log and simplifies locate's return type.

Split locate and from_longident code in two steps

- first identify the correct Uid by reducing the shape
- then load the cmt to find the corresponding location