Fix [locate] to look for non-pp source files. (#1219)
* Fix [locate] to look for non-pp source files.

When ppx is used, the cmt sourcefile that is recorded is the *.pp.mli? file,
and we compare its digest to the digests of the files in the source directory.
Often the .pp.mli? file isn't there, and even when it is we probably do not
want to prefer it over the normal .mli? file for locate. This patch checks if
the file is a .pp.mli? file, and if so uses the digest of the regular .mli?
file instead.

When sourcefile is a path it isn't a valid basename for
Misc.exact_file_exists. This feature fixes it, which resolves
test failures on MacOS.

* Adds a test analogous to dot-pp-dot-ml.t that builds with dune and
demonstrates that the fix works.

In order for the test to work we had to unset BUILD_PATH_PREFIX_MAP=

When this variable is set, $TESTCASE_ROOT appears in the build directory, which
causes the search for the source file to fail because the path does not contain
"$TESTCASE_ROOT" literally.

Co-authored-by: Ulysse Gérard <thevoodoos@gmail.com>