Windows: correctly redirect stderr to stdout of called ppx
Redirect the stdout of the child process to which of the parent (as
posix system(3)), and the stderr of the child to the stdout of the
child (as merlin's ppx_commandline).

Setting up the redirection in the parent process has the downside that
we must set the bInheritHandles flag of CreateProcess, which means
that if merlin doesn't set a handle non-inheritable (close-on-exec),
the ppx process will have access to it.

Reorder a bit the logging to print the parameters the ppx is called
with, and use an Unicode environment for the child process.