Add a cache mechanism for the PPX phase
This adds an all-or-nothing cache for the reader phase and for the PPX
phase; the important one being the latter.

The cache is disabled by default. It can be enabled via the new
USE_PPX_CACHE conifg directive, which is also added in this commit. The
build system should add the USE_PPX_CACHE directive if and only if the
project doesn't use PPX depnedencies.