Emacs: fix logging when merlin call is interrupted
Currently, "interrupted" is only logged when the result parsing is
interrupted (by C-g), but it seems pretty clear that the original
intent (and more useful behavior) was to log "interrupted" when the
actual merlin call was interrupted.

We do this by setting up separate condition handlers for (1) the
result parse failing; and (2) the merlin call or result parse being
interrupted.