Implement a Fuzzy CI to catch regressions
This commit implements a Fuzzy CI to catch end-to-end regressions in Merlin
by checking the responses of a set of ocamlmerlin queries. That set of
queries is determined by a deterministic but randomly chosen set of file
location samples on the Irmin code base. The CI passes automatically if the
responses on all samples remain the same before and after the PR. If there's a
diff in the responses, the CI fails and makes the diff available as a GH action
artifact.

There's a high-level description of the CI in much more detail on
https://github.com/ocaml/merlin/wiki/Merlin-Fuzzy-CI.

Co-authored-by: Enguerrand Decorne <decorne.en@gmail.com>