ocamlmerlin.c: improve location of socket path
On Windows, use [GetTempPath2][] if available or [GetTempPath][] to
retrieve the temporary directory path.

On other systems, follow the [XDG Base Directory Specification][XDG]
and put the socket file in the XDG runtime directory. It is indicated
by the `XDG_RUNTIME_DIR` environment variable, that is always set on
platforms implementing the spec. Otherwise, fallback to `TMPDIR` or
`/tmp`.

[GetTempPath2]: https://learn.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-gettemppath2a
[GetTempPath]: https://learn.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-gettemppatha
[XDG]: https://specifications.freedesktop.org/basedir/latest/