Improve passing arrays as parameters
In C, arrays passed as arguments to functions decay into pointers, and
the size information is lost. This means all of these are equivalent:

    void f(char *s);
    void f(char s[]);
    void f(char s[3615]);

One way to partially retain this information is to use `static n` with
`n` being the array size: the compiler now knows that the first `n`
elements after the pointed address can be accessed. This C99 feature
is unfortunately not supported by MSVC (but supported by clang-cl).
The compiler uses this info to check that the caller supplies a long
enough buffer to the callee. The size information is not used inside
the callee.