ocamlmerlin.c: use C99-conforming snprintf with mingw+msvcrt
_snprintf doesn't null-terminate strings.
MSVCRT doens't provide snprintf, but if __USE_MINGW_ANSI_STDIO is
defined, then mingw-w64 will provide this function.
There are environments where mingw-w64 is configured with MSVCRT, and
others where it is configured with UCRT. It seems safer to use the
compatibility shims.
UCRT provides snprintf, and MSVC defaults to UCRT.