Ensure CMSG accesses are aligned
To ensure the ancillary data buffer is suitably aligned, the
historical technique is to use a union with a backing char buffer and
a member having the desired alignment, as the alignment of the union
is the strictest alignment of its members. Another possibility is to
use the C11 alignas specifier. Keep the old technique for
compatibility.

Proper alignment should be ensured when creating the structure and
when reading it back from the kernel.

> `CMSG_DATA()`
>     returns a pointer to the data portion of a cmsghdr. The pointer
>     returned cannot be assumed to be suitably aligned for accessing
>     arbitrary payload data types. Applications should not cast it to
>     a pointer type matching the payload, but should instead use
>     `memcpy(3)` to copy data to or from a suitably declared object.

See also:
- https://man7.org/linux/man-pages/man7/unix.7.html
- https://man7.org/linux/man-pages/man3/cmsg.3.html
- https://pubs.opengroup.org/onlinepubs/9799919799/basedefs/sys_socket.h.html
- https://docs.kernel.org/core-api/unaligned-memory-access.html#why-unaligned-access-is-bad